Abstract API interfaces.
The HTTP API 0.6 is now just an implementation of abstract API
interfaces. Another implementation might some day use the native
database backend of OpenStreetMap.

The documentation has to be updated and restructured. Some
implementation classes still have to be fixed, everything needs to be
tested. Some code will be moved to the appropriate projects (Route
Manager and History Viewer).