Fixing exceptions after OSM API changes
Fixes FS#72

For deleted objects, the OSM API now returns an additional entry in the
version history that has the attribute visible="false". This entry does
not contain the usual data such as a lon and lat attribute for nodes. In
osmChange files, the <delete> elements now contain these non-visible
versions, where before they contained the (visible) version that was
there before the deletion.